Deleting some old code, keepin some others
I removed the old implementation of the Roomate class because it was too repetitive and the classes werent meaningful enough for the final project. I commented out Bill because love the way we have it above, might revisit because it makes our code more dynamic in terms of how many roomates there can be. My code is too hard coded so we can use this to fix that